tidy the PR and move the BD API part from env to spot.py
